requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0

# Testing
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3

# Additional Dependencies Found in Project
requests==2.31.0  # Used in test_admin_users.py
cryptography==41.0.7  # Required for JWT
bcrypt==4.1.2  # Required for password hashing